+++ tests/test_items.py
@@ -1,17 +1,6 @@
 import pytest
 
 from maimai_py import MaimaiClient
-from maimai_py.models import PlayerIdentifier
-from maimai_py.providers.arcade import ArcadeProvider
-
-
-@pytest.mark.asyncio(scope="session")
-async def test_regions(maimai: MaimaiClient, arcade: ArcadeProvider, arcade_player: PlayerIdentifier):
-    try:
-        regions = await maimai.regions(arcade_player, provider=arcade)
-        assert any(region.region_id == 2 for region in regions)
-    except Exception:
-        pytest.skip("Connection error, skipping the test.")
 
 
 @pytest.mark.asyncio(scope="session")